# cmd/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoFundIt is a collaborative platform enabling families and friends to plan and fund group activities, such as vacations and events. Features end-to-end encryption, secure payment handling, and flexible contribution options.",
        "title": "GoFundIt API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://gofundit.com/support",
            "email": "oyeniyibright@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/suggestions/activity": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Generates activity suggestions based on provided content text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suggestion"
                ],
                "summary": "Get Activity Suggestions by Text",
                "parameters": [
                    {
                        "description": "Content for suggestion generation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SuggestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity suggestions retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ActivitySuggestion"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid inputs",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/suggestions/activity/{campaignID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of suggested activities for a specific campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suggestion"
                ],
                "summary": "Get Campaign Activity Suggestions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity suggestions retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ActivitySuggestion"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            }
        },
        "/ws/campaign/{campaignID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Establishes a WebSocket connection for real-time updates about campaign activities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "websocket"
                ],
                "summary": "Campaign WebSocket Connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid campaign ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.SuggestionRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "description": "Content for generating suggestions",
                    "type": "string",
                    "example": "Trip to London for a day at midnight"
                }
            }
        },
...
        "handlers.BadRequestResponse": {
            "description": "Bad request error response structure",
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Validation errors (optional)\n@example [{\"field\":\"field\",\"error\":\"must be a valid data\"}]"
                },
                "message": {
                    "description": "Error message\n@example Invalid input provided",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"Bad Request\"\n@example Bad Request",
                    "type": "string",
                    "enum": [
                        "Bad Request"
                    ],
                    "example": "Bad Request"
                }
            }
        },
        "handlers.SuccessResponse": {
            "description": "Successful API response structure",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)\n@example null"
                },
                "message": {
                    "description": "Response message describing the success\n@example Operation completed successfully",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"OK\"\n@example OK",
                    "type": "string",
                    "enum": [
                        "OK"
                    ],
                    "example": "OK"
                }
            }
        },
        "handlers.UnauthorizedResponse": {
            "description": "Unauthorized error response structure",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Will always be \"Unauthorized\"\n@example Unauthorized",
                    "type": "string",
                    "enum": [
                        "Unauthorized"
                    ],
                    "example": "Unauthorized"
                },
                "status": {
                    "description": "Status will always be false for unauthorized\n@example false",
                    "type": "string",
                    "enum": [
                        "Unauthorized"
                    ],
                    "example": "Unauthorized"
                }
            }
        },
        "handlers.ValidationError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "must be a valid field"
                },
                "field": {
                    "type": "string",
                    "example": "field"
                }
            }
        },
        "handlers.response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)"
                },
                "errors": {
                    "description": "Error details (optional)"
                },
                "message": {
                    "description": "Response message\n@example Operation completed successfully",
                    "type": "string"
                },
                "status": {
                    "description": "Response status\n@example OK",
                    "type": "string"
                }
            }
        },
...
        "models.ActivitySuggestion": {
            "type": "object",
            "properties": {
                "activityName": {
                    "type": "string"
                },
                "estimatedCost": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for accessing the API endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token for authenticated requests. Use the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CampaignKeyAuth": {
            "description": "Campaign Key required for campaign-specific operations",
            "type": "apiKey",
            "name": "Campaign-Key",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
...
}